+++ R/app_ui.R
@@ -39,13 +39,17 @@ app_ui <- function(request) {
           ),
           a(
             # href = "https://github.com/CSSEGISandData/COVID-19",
-            href = "https://github.com/bumbeishvili/covid19-daily-data",
+            #href = "https://github.com/bumbeishvili/covid19-daily-data",
+            href =  "https://covid19datahub.io/",
             target = "_blank",
             span(
               id = "subtitle",
-              "Data source: worldometers from 26.03.2020, JHU CSSE before.",
+              #"Data source: worldometers from 26.03.2020, JHU CSSE before.",
+              "Data source: COVID-19 Data Hub",
               textOutput("last_update", inline = TRUE) %>%
-                bs_embed_tooltip(title = "Data Repository by bumbeishvili. More information on the README.", placement = "right")
+                #bs_embed_tooltip(title = "Data Repository by bumbeishvili. More information on the README.", placement = "right")
+                bs_embed_tooltip(title = "Data Repository by COVID-19 Data Hub. More information in the README on our github page.", placement = "right")
+
             )
           )
         ), # end header-left