+++ R/mod_caseBoxes.R
@@ -39,33 +39,33 @@ mod_caseBoxes_server <- function(input, output, session, counts) {
 
   output$confirmed <- renderUI({
     countBox(title1 = "Confirmed: ",
-             subtitle1 = counts()[["confirmed"]],
+             subtitle1 = counts[["confirmed"]],
              title2 = "New: ",
-             subtitle2 =  counts()[["new_confirmed"]],
+             subtitle2 =  counts[["new_confirmed"]],
              color = "white",
              background = case_colors[["confirmed"]])
   })
   output$death <- renderUI({
     countBox(title1 = "Deaths: ",
-             subtitle1 = counts()[["deaths"]],
+             subtitle1 = counts[["deaths"]],
              title2 = "New: ",
-             subtitle2 =  counts()[["new_deaths"]],
+             subtitle2 =  counts[["new_deaths"]],
              color = "white",
              background = case_colors[["deaths"]])
   })
   output$recovered <- renderUI({
     countBox(title1 = "Recovered: ",
-             subtitle1 = counts()[["recovered"]],
+             subtitle1 = counts[["recovered"]],
              title2 = "New: ",
-             subtitle2 =  counts()[["new_recovered"]],
+             subtitle2 =  counts[["new_recovered"]],
              color = "white",
              background = case_colors[["recovered"]])
   })
   output$active <- renderUI({
     countBox(title1 = "Active: ",
-             subtitle1 = counts()[["active"]],
+             subtitle1 = counts[["active"]],
              title2 = "New: ",
-             subtitle2 =  counts()[["new_active"]],
+             subtitle2 =  counts[["new_active"]],
              color = "white",
              background = case_colors[["active"]])
   })