+++ R/mod_continent_comparison.R
@@ -39,8 +39,7 @@ mod_continent_comparison_ui <- function(id){
 
 #' continent_comparison Server Function
 #'
-#' @param orig_data_aggregate reactive data.frame
-#' @param countries reactive data.frame
+#' @param orig_data_aggregate data.frame
 #' @param n min number of cases for a country to be considered. Default 1000
 #' @param w number of days of outbreak. Default 7
 #'
@@ -50,9 +49,8 @@ mod_continent_comparison_ui <- function(id){
 mod_continent_comparison_server <- function(input, output, session, orig_data_aggregate, n = 1000, w = 7, pop_data){
   ns <- session$ns
 
-  statuses <- c("confirmed", "deaths", "recovered", "active")
   # select all variables
-  allstatuses = c(statuses, paste0("new_", statuses))
+  #allstatuses = get_aggrvars()
 
   #continents = reactive({unique(orig_data_aggregate()$continent)})
 
@@ -60,13 +58,19 @@ mod_continent_comparison_server <- function(input, output, session, orig_data_ag
   #   group_by(continent) %>%
   #   summarize(population = sum(as.numeric(population), na.rm = T))
   # aggregate data to continent
-  continent_data <- reactive({aggr_to_cont(orig_data_aggregate(), "continent", "date", pop_data, allstatuses)})
-  continents = reactive({unique(continent_data()$Country.Region)})
+  continent_data <- #reactive({
+    aggr_to_cont(orig_data_aggregate, "continent", "date", #pop_data,
+                                           #allstatuses
+                                           )
+  #})
+  continents = #reactive({
+    unique(continent_data$Country.Region)
+  #})
 
-  #continent_data_filtered <- reactive({aggr_to_cont(data_filtered(), "continent", "date", pop_data, allstatuses)})
-  continent_data_filtered <- reactive({continent_data() %>% # select continents with longer outbreaks
+  continent_data_filtered <- #reactive({
+    continent_data %>% # select continents with longer outbreaks
       rescale_df_contagion(n = n, w = w)
-      })
+      #})
 
 
   output$lineplots_cont <- renderUI({
@@ -83,7 +87,7 @@ mod_continent_comparison_server <- function(input, output, session, orig_data_ag
     mod_growth_death_rate_ui(ns("rate_plots_cont"), n_highligth = length(continents()))
   })
 
-  callModule(mod_growth_death_rate_server, "rate_plots_cont", continent_data_filtered, n = n, n_highligth = length(continents()), istop = F)
+  callModule(mod_growth_death_rate_server, "rate_plots_cont", continent_data_filtered, n = n, n_highligth = length(continents), istop = F)
 
   # Line with bullet plot
 
@@ -92,20 +96,20 @@ mod_continent_comparison_server <- function(input, output, session, orig_data_ag
   })
 
   callModule(mod_compare_nth_cases_plot_server, "lines_points_plots_cont", continent_data_filtered, n = n, w = w,
-             n_highligth = length(continents()), istop = F)
+             n_highligth = length(continents), istop = F)
 
   # scatterplot
   output$scatterplot_plots_cont <- renderUI({
     mod_scatterplot_ui(ns("scatterplot_plots_cont"))
   })
 
-  callModule(mod_scatterplot_server, "scatterplot_plots_cont", continent_data_filtered, n = n, n_highligth = length(continents()), istop = F, countries = continents)
+  callModule(mod_scatterplot_server, "scatterplot_plots_cont", continent_data_filtered, n = n, n_highligth = length(continents), istop = F, countries = continents)
 
 
   output$status_stackedbarplot_cont <- renderUI({
     mod_stackedbarplot_ui(ns("status_stackedbarplot_cont"))
   })
-  callModule(mod_stackedbarplot_status_server, "status_stackedbarplot_cont", continent_data_filtered, n = n, n_highligth = length(continents()), istop = F)
+  callModule(mod_stackedbarplot_status_server, "status_stackedbarplot_cont", continent_data_filtered, n = n, n_highligth = length(continents), istop = F)
 
   # tables ----
   callModule(mod_add_table_server, "add_table_cont", continent_data_filtered, maxrowsperpage = 10)